requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-ollama
langchain-groq
langchain
faiss-cpu